Move compute_chrf() to evaluation module per review feedback
- Created new src/muse/evaluation module for MT evaluation metrics
- Moved compute_chrf() from src/muse/translation/metrics.py to src/muse/evaluation/metrics.py
- Updated test script import path
- Added evaluation to module exports

This separates evaluation concerns from translation generation, allowing for future expansion with other metrics (COMET, BLEU, etc.)
