+++ src/muse/evaluation/__init__.py
@@ -1,6 +1,2 @@
 # Copyright (c) 2025, Center for Digital Humanities, Princeton University
 # SPDX-License-Identifier: Apache-2.0
-
-__version__ = "0.1.dev1"
-
-__all__ = ["__version__", "evaluation", "parallel_corpus", "translation"]
